proj3/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
install==1.3.5
protobuf==4.23.2
